pyproject.toml
[tool.poetry]
name = "Challenge 5"
version = "0.1.0"
description = "My solutions to the fifth challenge"
authors = ["luca-nik <luca.nicoli.engineer@gmail.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.3"
pandas = "^2.2.3"
pyarrow = "^18.0.0"
matplotlib = "^3.9.2"
mplfinance = "^0.12.10-beta.0"
seaborn = "^0.13.2"
python-dotenv = "^1.0.1"
torch = "^2.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
